+++ CHANGES.md
@@ -5,7 +5,8 @@ Changes in 8.7.1
 ----------------
 - Fix proc_open() PHP function not allowed in PDF.php
 - Fix PHP Warning A non-numeric value encountered in ReportCards.fnc.php
-- Fix PHP Fatal error Unsupported operand types in Teacher Programs: do not search Students List, unset in CustomFields.fnc.php & GetStaffList.fnc.php
+- Fix PHP Fatal error Unsupported operand types in Teacher Programs: do not search Students List, unset in CustomFields.fnc.php
+- Add 'staff_' prefix to first & last inputs on Find a User form in GetStaffList.fnc.php & Search.fnc.php
 
 Changes in 8.7
 --------------